@@ -17,7 +17,8 @@
     ],
     "require": {
         "php": ">=5.3.9",
-        "symfony/polyfill-mbstring": "~1.0"
+        "symfony/polyfill-mbstring": "~1.0",
+        "symfony/debug": "~2.7,>=2.7.2|~3.0.0"
     },
     "require-dev": {
         "symfony/event-dispatcher": "~2.1|~3.0.0",
