@@ -19,7 +19,7 @@
         "php": ">=5.3.9"
     },
     "require-dev": {
-        "symfony/yaml": "~2.1|~3.0.0",
+        "symfony/yaml": "~2.3.42|~2.7.14|~2.8.7|~3.0.7",
         "symfony/config": "~2.2|~3.0.0",
         "symfony/expression-language": "~2.6|~3.0.0"
     },
@@ -29,6 +29,7 @@
     "suggest": {
         "symfony/yaml": "",
         "symfony/config": "",
+        "symfony/expression-language": "For using expressions in service container configuration",
         "symfony/proxy-manager-bridge": "Generate service proxies to lazy load them"
     },
     "autoload": {
