@@ -18,7 +18,7 @@
     "require": {
         "php": ">=5.3.9",
         "symfony/event-dispatcher": "~2.6,>=2.6.7|~3.0.0",
-        "symfony/http-foundation": "~2.5,>=2.5.4|~3.0.0",
+        "symfony/http-foundation": "~2.7.20|~2.8.13|~3.1.6",
         "symfony/debug": "~2.6,>=2.6.2",
         "psr/log": "~1.0"
     },
