@@ -17,10 +17,11 @@
     ],
     "require": {
         "php": ">=5.3.9",
+        "symfony/polyfill-mbstring": "~1.0",
         "symfony/translation": "~2.4|~3.0.0"
     },
     "require-dev": {
-        "symfony/http-foundation": "~2.1|~3.0.0",
+        "symfony/http-foundation": "~2.3|~3.0.0",
         "symfony/intl": "~2.7.4|~2.8|~3.0.0",
         "symfony/yaml": "~2.0,>=2.0.5|~3.0.0",
         "symfony/config": "~2.2|~3.0.0",
