@@ -1031,8 +1031,17 @@ class DrupalKernel implements DrupalKernelInterface, TerminableInterface {
       $prefix = Settings::getApcuPrefix('class_loader', $this->root);
       $apc_loader = new ApcClassLoader($prefix, $this->classLoader);
       $this->classLoader->unregister();
-      $apc_loader->register();
+      // The optimized classloader might be persistent and store cache misses.
+      // For example, once a cache miss is stored in APCu clearing it on a
+      // specific web-head will not clear any other web-heads. Therefore
+      // fallback to the composer class loader that only statically caches
+      // misses.
+      $old_loader = $this->classLoader;
       $this->classLoader = $apc_loader;
+      // Our class loaders are preprended to ensure they come first like the
+      // class loader they are replacing.
+      $old_loader->register(TRUE);
+      $apc_loader->register(TRUE);
     }
   }
 
