@@ -100,11 +100,12 @@
         $(window)
           .on('resize.dialogResize scroll.dialogResize', eventData, autoResize)
           .trigger('resize.dialogResize');
-        $(document).on('drupalViewportOffsetChange', eventData, autoResize);
+        $(document).on('drupalViewportOffsetChange.dialogResize', eventData, autoResize);
       }
     },
     'dialog:beforeclose': function (event, dialog, $element) {
       $(window).off('.dialogResize');
+      $(document).off('.dialogResize');
     }
   });
 
