@@ -648,7 +648,7 @@ class ConfigTranslationUiTest extends WebTestBase {
 
     foreach ($languages as $langcode => $data) {
       // Import a .po file to add a new language with a given number of plural forms
-      $name = tempnam('temporary://', $langcode . '_') . '.po';
+      $name = \Drupal::service('file_system')->tempnam('temporary://', $langcode . '_') . '.po';
       file_put_contents($name, $this->getPoFile($data['plurals']));
       $this->drupalPostForm('admin/config/regional/translate/import', array(
         'langcode' => $langcode,
@@ -684,7 +684,7 @@ class ConfigTranslationUiTest extends WebTestBase {
 
     // First import a .po file with multiple plural forms.
     // This will also automatically add the 'sl' language.
-    $name = tempnam('temporary://', "sl_") . '.po';
+    $name = \Drupal::service('file_system')->tempnam('temporary://', "sl_") . '.po';
     file_put_contents($name, $this->getPoFile(4));
     $this->drupalPostForm('admin/config/regional/translate/import', array(
       'langcode' => 'sl',
