@@ -4485,6 +4485,18 @@ $connection->insert('field_data_field_file')
   'field_file_display' => '1',
   'field_file_description' => 'file desc',
 ))
+->values(array(
+  'entity_type' => 'user',
+  'bundle' => 'user',
+  'deleted' => '0',
+  'entity_id' => '2',
+  'revision_id' => '2',
+  'language' => 'und',
+  'delta' => '0',
+  'field_file_fid' => '2',
+  'field_file_display' => '1',
+  'field_file_description' => 'file desc',
+))
 ->execute();
 
 $connection->schema()->createTable('field_data_field_float', array(
@@ -6674,6 +6686,18 @@ $connection->insert('field_revision_field_file')
   'field_file_display' => '1',
   'field_file_description' => 'file desc',
 ))
+->values(array(
+  'entity_type' => 'user',
+  'bundle' => 'user',
+  'deleted' => '0',
+  'entity_id' => '2',
+  'revision_id' => '2',
+  'language' => 'und',
+  'delta' => '0',
+  'field_file_fid' => '2',
+  'field_file_display' => '1',
+  'field_file_description' => 'file desc',
+))
 ->execute();
 
 $connection->schema()->createTable('field_revision_field_float', array(
@@ -8307,6 +8331,16 @@ $connection->insert('file_managed')
   'status' => '1',
   'timestamp' => '1421727515',
 ))
+->values(array(
+  'fid' => '2',
+  'uid' => '1',
+  'filename' => 'ds9.txt',
+  'uri' => 'public://ds9.txt',
+  'filemime' => 'text/plain',
+  'filesize' => '4720',
+  'status' => '1',
+  'timestamp' => '1421727516',
+))
 ->execute();
 
 $connection->schema()->createTable('file_usage', array(
@@ -8373,6 +8407,13 @@ $connection->insert('file_usage')
   'module' => 'file',
   'type' => 'node',
   'id' => '1',
+  'count' => '1',
+))
+->values(array(
+  'fid' => '2',
+  'module' => 'file',
+  'type' => 'user',
+  'id' => '2',
   'count' => '1',
 ))
 ->execute();
@@ -43860,7 +43901,7 @@ $connection->insert('variable')
 ))
 ->values(array(
   'name' => 'menu_override_parent_selector',
-  'value' => 'b:0;',
+  'value' => 'b:1;',
 ))
 ->values(array(
   'name' => 'menu_parent_article',
