@@ -73,6 +73,11 @@ class NodePreviewForm extends FormBase {
     $view_mode = $node->preview_view_mode;
 
     $query_options = array('query' => array('uuid' => $node->uuid()));
+    $query = $this->getRequest()->query;
+    if ($query->has('destination')) {
+      $query_options['query']['destination'] = $query->get('destination');
+    }
+
     $form['backlink'] = array(
       '#type' => 'link',
       '#title' => $this->t('Back to content editing'),
@@ -80,9 +85,11 @@ class NodePreviewForm extends FormBase {
       '#options' => array('attributes' => array('class' => array('node-preview-backlink'))) + $query_options,
     );
 
-    $view_mode_options = $this->entityManager->getViewModeOptionsByBundle('node', $node->bundle());
+    // Always show full as an option, even if the display is not enabled.
+    $view_mode_options = ['full' => $this->t('Full')] + $this->entityManager->getViewModeOptionsByBundle('node', $node->bundle());
 
     // Unset view modes that are not used in the front end.
+    unset($view_mode_options['default']);
     unset($view_mode_options['rss']);
     unset($view_mode_options['search_index']);
 
@@ -116,10 +123,18 @@ class NodePreviewForm extends FormBase {
    * {@inheritdoc}
    */
   public function submitForm(array &$form, FormStateInterface $form_state) {
-    $form_state->setRedirect('entity.node.preview', array(
+    $route_parameters = [
       'node_preview' => $form_state->getValue('uuid'),
       'view_mode_id' => $form_state->getValue('view_mode'),
-    ));
+    ];
+
+    $options = [];
+    $query = $this->getRequest()->query;
+    if ($query->has('destination')) {
+      $options['query']['destination'] = $query->get('destination');
+      $query->remove('destination');
+    }
+    $form_state->setRedirect('entity.node.preview', $route_parameters, $options);
   }
 
 }
