@@ -22,6 +22,9 @@ class NodeForm extends ContentEntityForm {
 
   /**
    * Whether this node has been previewed or not.
+   *
+   * @deprecated Scheduled for removal in Drupal 8.3.x. Use the form state
+   *   property 'has_been_previewed' instead.
    */
   protected $hasBeenPreviewed = FALSE;
 
@@ -65,6 +68,8 @@ class NodeForm extends ContentEntityForm {
    * {@inheritdoc}
    */
   public function form(array $form, FormStateInterface $form_state) {
+    $this->hasBeenPreviewed = $form_state->get('has_been_previewed') ?: FALSE;
+
     // Try to restore from temp store, this must be done before calling
     // parent::form().
     $store = $this->tempStoreFactory->get('node_preview');
@@ -89,6 +94,7 @@ class NodeForm extends ContentEntityForm {
       $this->entity = $preview->getFormObject()->getEntity();
       $this->entity->in_preview = NULL;
 
+      $form_state->set('has_been_previewed', TRUE);
       $this->hasBeenPreviewed = TRUE;
     }
 
@@ -231,7 +237,7 @@ class NodeForm extends ContentEntityForm {
     $node = $this->entity;
     $preview_mode = $node->type->entity->getPreviewMode();
 
-    $element['submit']['#access'] = $preview_mode != DRUPAL_REQUIRED || $this->hasBeenPreviewed;
+    $element['submit']['#access'] = $preview_mode != DRUPAL_REQUIRED || $form_state->get('has_been_previewed');
 
     // If saving is an option, privileged users get dedicated form submit
     // buttons to adjust the publishing status while saving in one go.
@@ -338,10 +344,19 @@ class NodeForm extends ContentEntityForm {
     $store = $this->tempStoreFactory->get('node_preview');
     $this->entity->in_preview = TRUE;
     $store->set($this->entity->uuid(), $form_state);
-    $form_state->setRedirect('entity.node.preview', array(
+
+    $route_parameters = [
       'node_preview' => $this->entity->uuid(),
-      'view_mode_id' => 'default',
-    ));
+      'view_mode_id' => 'full',
+    ];
+
+    $options = [];
+    $query = $this->getRequest()->query;
+    if ($query->has('destination')) {
+      $options['query']['destination'] = $query->get('destination');
+      $query->remove('destination');
+    }
+    $form_state->setRedirect('entity.node.preview', $route_parameters, $options);
   }
 
   /**
