@@ -59,13 +59,21 @@ class UrlAlias extends DestinationBase implements ContainerFactoryPluginInterfac
    * {@inheritdoc}
    */
   public function import(Row $row, array $old_destination_id_values = array()) {
+    $source = $row->getDestinationProperty('source');
+    $alias = $row->getDestinationProperty('alias');
+    $langcode = $row->getDestinationProperty('langcode');
+    $pid = $old_destination_id_values ? $old_destination_id_values[0] : NULL;
 
-    $path = $this->aliasStorage->save(
-      $row->getDestinationProperty('source'),
-      $row->getDestinationProperty('alias'),
-      $row->getDestinationProperty('langcode'),
-      $old_destination_id_values ? $old_destination_id_values[0] : NULL
-    );
+    // Check if this alias is for a node and if that node is a translation.
+    if (preg_match('/^\/node\/\d+$/', $source) && $row->hasDestinationProperty('node_translation')) {
+
+      // Replace the alias source with the translation source path.
+      $node_translation = $row->getDestinationProperty('node_translation');
+      $source = '/node/' . $node_translation[0];
+      $langcode = $node_translation[1];
+    }
+
+    $path = $this->aliasStorage->save($source, $alias, $langcode, $pid);
 
     return array($path['pid']);
   }
