@@ -240,7 +240,10 @@ abstract class TestBase {
   /**
    * The temporary file directory for the test environment.
    *
-   * This is set in TestBase::prepareEnvironment().
+   * This is set in TestBase::prepareEnvironment(). This value has to match the
+   * temporary directory created in install_base_system() for test installs.
+   *
+   * @see install_base_system()
    *
    * @var string
    */
