@@ -113,7 +113,12 @@ class UncaughtExceptionTest extends WebTestBase {
    * Tests a missing dependency on a service.
    */
   public function testMissingDependency() {
-    $this->expectedExceptionMessage = 'Argument 1 passed to Drupal\error_service_test\LonelyMonkeyClass::__construct() must be an instance of Drupal\Core\Database\Connection, non';
+    if (version_compare(PHP_VERSION, '7.1') < 0) {
+      $this->expectedExceptionMessage = 'Argument 1 passed to Drupal\error_service_test\LonelyMonkeyClass::__construct() must be an instance of Drupal\Core\Database\Connection, non';
+    }
+    else {
+      $this->expectedExceptionMessage = 'Too few arguments to function Drupal\error_service_test\LonelyMonkeyClass::__construct(), 0 passed';
+    }
     $this->drupalGet('broken-service-class');
     $this->assertResponse(500);
 
