@@ -120,7 +120,7 @@ class TermTest extends TaxonomyTestBase {
     // Set limit to 10 terms per page. Set variable to 9 so 10 terms appear.
     $this->config('taxonomy.settings')->set('terms_per_page_admin', '9')->save();
     $term1 = $this->createTerm($this->vocabulary);
-    $terms_array = '';
+    $terms_array = [];
 
     $taxonomy_storage = $this->container->get('entity.manager')->getStorage('taxonomy_term');
 
