@@ -9,7 +9,7 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
 /**
  * Provides block plugin definitions for all Views exposed filters.
  *
- * @see \Drupal\views\Plugin\block\block\ViewsExposedFilterBlock
+ * @see \Drupal\views\Plugin\Block\ViewsExposedFilterBlock
  */
 class ViewsExposedFilterBlock implements ContainerDeriverInterface {
 
